Make field validator class with static functions a service for easier unit testing
